Change project

+++ model/add_news_db.php
@@ -31,15 +31,30 @@
                 'news_body' => $v['body_news'],
                 'img_puth' => $v['image_puth'],
                 'user_pub_id' => $v['offer_user_id']]);
+           $lastInsertId = $conn->lastInsertId();
            if($result)
            {
-               $delete_news = $conn->prepare('DELETE FROM tmp_news WHERE id=?');
-               if($delete_news->execute([$v['id']]))
-                   return true;
-               else
+              $update_table_tems = $conn->prepare('UPDATE tems SET  news_id=? WHERE tmp_news_id=?');
+              if($update_table_tems->execute([$lastInsertId, $v['id']])){
+                $delete_news = $conn->prepare('DELETE FROM tmp_news WHERE id=?');
+                if($delete_news->execute([$v['id']]))
+                  return true;
+              }else
                    return false;
            }else
                return false;
 
         }
+    }
+
+    function addAdminNews($name_news, $body, $tema, $user_id, $image_puth)
+    {
+        $conn = con();
+        $sql = $conn->prepare("INSERT INTO news (news_name, news_body, image_puth, user_id) VALUES (?, ?, ?, ?)");
+        if($sql->execute([$name_news, $body, $image_puth, $user_id])){
+          $lastInsertId = $conn->lastInsertId();
+          $sql = $conn->prepare('INSERT INTO tems (tema, news_id) VALUES (?, ?)');
+          if($sql->execute([$tema, $lastInsertId]))
+            return true;
+        }else return false;
     }
